pyproject.toml
[project]
name = "fails"
version = "0.1.0"
description = "Generate insights from W&B Weave Evals"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "litellm>=1.73.1",
    "openai-agents[litellm]>=0.0.19",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
    "simple-parsing>=0.1.7",
    "weave>=0.51.56",
    "scikit-learn>=1.3.0",
    "scipy>=1.9.0",
]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["fails"]
